Added support for decimal point calculations. Also, output is now displayed up to 3 decimal places, rather than just being rounded to the nearest whole Integer.

+++ src/main/java/com/example/calculator/operators.java
@@ -23,22 +23,22 @@ public class operators {
 
         switch (x) {
             case '+' -> {
-                Controller.total = Integer.parseInt(modifyDisplay.getDisplayText());
+                Controller.total = Double.parseDouble(modifyDisplay.getDisplayText());
                 modifyDisplay.clearDisplay();
                 flag = '+';
             }
             case '-' -> {
-                Controller.total = Integer.parseInt(modifyDisplay.getDisplayText());
+                Controller.total = Double.parseDouble(modifyDisplay.getDisplayText());
                 modifyDisplay.clearDisplay();
                 flag = '-';
             }
             case '/' -> {
-                Controller.total = Integer.parseInt(modifyDisplay.getDisplayText());
+                Controller.total = Double.parseDouble(modifyDisplay.getDisplayText());
                 modifyDisplay.clearDisplay();
                 flag = '/';
             }
             case '*' -> {
-                Controller.total = Integer.parseInt(modifyDisplay.getDisplayText());
+                Controller.total = Double.parseDouble(modifyDisplay.getDisplayText());
                 modifyDisplay.clearDisplay();
                 flag = '*';
             }
@@ -48,20 +48,19 @@ public class operators {
     public static void equals(){
         switch (flag) {
             case '+' -> {
-                Controller.total += Integer.parseInt(modifyDisplay.getDisplayText());
+                Controller.total += Double.parseDouble(modifyDisplay.getDisplayText());
                 modifyDisplay.setTextOperation();
-
             }
             case '-' -> {
-                Controller.total -= Integer.parseInt(modifyDisplay.getDisplayText());
+                Controller.total -= Double.parseDouble(modifyDisplay.getDisplayText());
                 modifyDisplay.setTextOperation();
             }
             case '/' -> {
-                Controller.total /= Integer.parseInt(modifyDisplay.getDisplayText());
+                Controller.total /= Double.parseDouble(modifyDisplay.getDisplayText());
                 modifyDisplay.setTextOperation();
             }
             case '*' -> {
-                Controller.total *= Integer.parseInt(modifyDisplay.getDisplayText());
+                Controller.total *= Double.parseDouble(modifyDisplay.getDisplayText());
                 modifyDisplay.setTextOperation();
             }
         }
